Lógica de reservas

+++ models/Locais.py
@@ -7,4 +7,4 @@ class Locais(db.Model):
     Id_Local = db.Column(db.Integer, primary_key=True, autoincrement=True)
     Endereco = db.Column(db.String(100), nullable = False)
     Nome = db.Column(db.String(50), nullable = False)
-    Porcentagem = db.Column(db.DECIMAL(10,3), nullable = False)
+    Porcentagem = db.Column(db.Float, nullable = False)
